Added spotify_is_active
Returns True when the user is considered to be active in their current
room, False otherwise. This is a pretty naive implementation at the
moment, as it won't account for track and progress_ms being at the end
of the previous song. There might be an easy heuristic to guard against
these edge cases, such as not dropping the user if progress_ms + time
elapsed since sync is greater than track duration. If they really have
dropped off then we'll pick that up in the next sync.